@ r moved Authorization to CosmoQuestX namespace - may have borked logins
