+++ csb/csb-accounts/auth.php
@@ -10,35 +10,6 @@
    Start / renew the session
    ---------------------------------------------------------------------- */
 session_start();
-class Authorization{
-
-    /**
-     * Compare the user name against the database for a given id
-     *
-     * @param resource $db
-     * @param int $id
-     * @param string $name
-     * @return boolean
-     */
-    public static function chk_UserId($db, $id, $name): bool
-    {
-
-        $query = "SELECT id, name FROM users WHERE id = ?";
-        $params = array($id);
-        $result = $db->runQueryWhere($query, "s", $params)[0];
-
-        // strip out any white space and make everything lower case because typing
-        $comp = strtolower(trim($result['name'], "\t\n\r\0\x0B"));
-        $name = strtolower(trim($name, " \t\n\r\0\x0B"));
-
-
-        if (!strcmp($comp, $name))
-            return TRUE;
-        else
-            return FALSE;
-
-    }
-}
 
 /**
  * Compare the "remember me" cookie against the database for a given token
@@ -79,7 +50,7 @@ function isLoggedIn($db)
 
     // look for the session id to be valid
     if (isset($_SESSION['user_id']) && !empty($_SESSION['user_id']) && !empty($_COOKIE["name"])) {
-        $flag = Authorization::chk_UserId($db, $_SESSION['user_id'], $_COOKIE["name"]);
+        $flag = CosmoQuestX\Authorization::chk_UserId($db, $_SESSION['user_id'], $_COOKIE["name"]);
     } // see if the cookie - WHICH CAN BE TAMPERED WITH - matches the DB
     elseif (!empty($_COOKIE["name"]) && !empty($_COOKIE["token"])) {
         $flag = chk_Token($db, $_COOKIE["token"], $_COOKIE["name"]);
